:recycle: Refactor: 파일 초기 세팅
