+++ feature/dailyHugg/src/main/java/com/hugg/dailyhugg/all/DailyHuggListPageState.kt
@@ -2,6 +2,6 @@ package com.hugg.dailyhugg.all
 
 import com.hugg.feature.base.PageState
 
-data class DailyHuggByRoundPageState(
-
+data class DailyHuggListPageState(
+    val temp: String = ""
 ): PageState
