+++ feature/dailyHugg/src/main/java/com/hugg/dailyhugg/all/DailyHuggListViewModel.kt
@@ -1,4 +1,12 @@
 package com.hugg.dailyhugg.all
 
-class DailyHuggByRoundViewModel {
+import com.hugg.feature.base.BaseViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class DailyHuggListViewModel @Inject constructor(
+
+) : BaseViewModel<DailyHuggListPageState>(DailyHuggListPageState()) {
+    
 }
